Output EFI GUIDs of protocols supported on the app image handle
This functionality is useful to explore what handles support what
protocols. I want to try to implement a simple logic of finding a file
in the same directory as our binary and load it in memory.

I assume that we should be able to use the handle passed to the entry
point to figure out the file path, which would allow us to find the
right directory path.

Also via the handle passed to the entry point we should be able to have
access to the volume where the file is located. With that we can use
UEFI File APIs to read the file in memory.

The fundamental idea of this exercesize is that loading and OS image is
basically loading file in memory and the passing control to the right
place in memory. So it should be a simple enough demostration of how an
EFI based OS loader could work.