Remove remains of the test binary from Makefile
I used a separate binary to test the C-library-like utils I created, but
it was really awkward and I think there are better means to go around
unit tests in C, so I didn't push it to the repository, but didn't fix
the build script. This change fixes the build script.