Introduce EFI Simple File System and File protocol definitions
As the names of the protocols suggest they are needed to manipulate
files. I introduced only a subset of all definitions that I want to use
in the future to load image in memory.

Along the way I updated the EFI Boot Services table to also include the
function to dynamically allocate memory:

 * struct efi_file_info kind of implies dynamic memory allocation,
   though I want to avoid it to make the code slightly simpler
 * we need to allocate memory to read the buffer into, though in the end
   it's probably better to use page allocation methods instead of pool
   allocation methods for the image.