Add function to get/set the current position in the file
With the currently defined function we can only read the file from the
beginning to the end sequentially, which is not particularly convenient
when we need to read structured binary files.

This changes fixes it by introducing functions that get/set the current
position in the file, so we can set the file offset before reading.