Add an example of reading and parsing ELF header
In the UNIX world the ELF file format is more widespread then the
Windows PE32+ format used for UEFI applications. So in the future I want
to be able to use this loader to load ELF files.

This change introduces basic definitions for ELF file format and a dummy
ELF binary to test the code. It also changes the code to read and dump
ELF header.