Code reading and parsing ELF program headers
ELF program headers is what we mostly care about when we want to load a
program in memory, so reading and understanding program headers is the
next small step towards loading ELF executable in memory.

Before reading and parsing program headers from the ELF file it makes
sense to verify the ELF file itself to make sure that it's an ELF file
to begin with, but besides that that it's an ELF file that we can
actually support.

ELF file format supports multiple different platforms and types of
files. We don't care about many of them at this point (like we
definitely don't care about object files, also I don't want to support
32-bit ELF files, etc).

To report verification errors I had to introduce some EFI common error
codes. And since I introduced them, along the way I replaced all the
status related magic constants with the appropriate constants.