Return a non-zero result from the main in kernel.c
Returning a non-zero result will help us to test if transfered control
to the loaded ELF image correctly and got control back or not by testing
the returned value.

The returned value coincidentally can be 42, but it's somewhat unlikely,
but just in case we can double check by changing the value to something
else in the future.