Add code to load a position independ ELF image and transfer control
This code assumes that the ELF image is position independent and doesn't
require dynamic linking (so no GOT or PLT manipulations are required).

The change additionally assumes that the ELF image will return the
control back, which is there solely for the purpose of testing and will
be removed later.

The testing is done by checking the returned value against a predermined
magic number. The images used for testing must all return 42.