Support building all the targets for aarch64 (64-bit ARM)
UEFI Specification also supports ARM. I have a HiKey960 board that uses
UEFI firmware. I want to try all of this on that board.

The aarch64 bit support is a bit ad-hoc. Basically it boils down to
changing compilation flags depending on the target achritecture.
Architecture specific flags live in the corresponding .env file
(x86-64.env or aarch64.env).

The right file is included into the main file depending on the value of
the ARCH variable, which is by default x86-64. If you want to compile
for aarch64 you can run:

$ ARCH=aarch64 make

Additionally, aarch64 target doesn't support the sysv_abi attrbute, so
I had to wrap it into a preprocessor macro. It appears that Microsoft
ABI for aarch64 basically the same as the GNU ABI.

Finally, to keep the main Makefile structure as uniform as possible I
changed the name of the output files to be arch-independent.

I tested the change with the QEMU and UEFI implementation for QEMU and
it appears to work fine. So it should be ready to be tested on the real
hardware.