Exit boot services before calling into ELF binary
We probably could live without that, but exiting EFI boot services
before transfering control to the loaded kernel should be a part of
normal control transfer since EFI will not be controlling memory
anymore.

To exit boot services we need to provide EFI with the memory map key.
The key can be obtained from the get_memory_map call. We need to keep in
mind however that whatever memory we need to allocate to hold the memory
map might change that memory map and therefore the memory map key.
Similarly freeing allocated memory might change the memory map and the
memory map key.

As a result we cannot really free the memory we allocate for the memory
map before exiting boot services, so this memory is intentionally
leaked.

This leak however doesn't matter much, since it should not be a big
amount of memory anyways. Plus the kernel of the operating system can
use whatever memory it sees fit, so we don't need to worry that this
memory could not be used anymore.

Also, the proper bootloader would probably pass the memory map to the
kernel somehow, since it's kind of an important piece of information for
the kernel to work.

Along the way I fixed formatting a little bit.