Do not return from the kernel.c main function
ELF loader code used to wait for the main to return to check if it was
actually called. That was a useful mode of operation for manual testing.
Now when the testing is done, I don't need it anymore and the ELF loader
code doesn't actually expect the loaded binary to return.