Introduce the logic to read a configuration file
At this point this logic isn't used anywhere and from the configuration
this part only has name and that's it (this configuration is not being
parsed or used in any way). Eventually though I'd like to be able to
configure the loader with a simple text file.

The point of having a configuration text file is that I want to be able
to read, load and pass to the kernel a Device Tree blob to the loaded
kernel. The general idea is that in the configuration file I will
specify the files that need to be read and loaded in memory (the kernel
binary itself, the Device Tree blob and whatsnot).

Also since the code is getting way to complicated and involved for just
one main.c, I'm starting to extract part of the logic into spearate
files.

Finally, if the loader is not a continiously running binary, so I'm
saving a bit of effort on the cleanup code in case of failure and do not
try to free all the memory loader allocated. If it turns out to be a
problem it should be relatively straighforward to add tracking of the
allocated memory and then free it all at the end in case of failures.