Add logic to parse loader configuration file
The format of the configuration file is rather simplistic. It will
contain of key value pairs. Keys will be separated from values by ':'
character. And key:value pairs will be seprated by any kind of a space
character. Additional spaces are allowed.

Each key:value pair will describe a module that need to be loaded in
memory. Among the modules one must be named "kernel" and point to an ELF
binary that will be loaded by the loader as an EFL binary. The loader
will then pass the control to the loaded ELF binary.

Other modules will be simple data modules and they will be loaded in
memory as-is, unlike the ELF files. The information about all the
modules will be passed to the ELF entry point function as a parameter.

This code was only tested to compile for x86 and aarch64, I didn't
actually test that it works correctly.