Add an example config file
In this example config file the efi\boot\kernel files is loaded twice.
First time as ELF binary that the loader will pass control to. And
additionally the same file is loaded as data that will be passed to the
loaded kernel.

There isn't terribly a lot of sense in loading the same file twice, it's
just an example of how we can pass things to the kernel using this
mechanism.