Update EFI loader to support loading modules
This change combines together a few of the previous changes and
completes the logic inside loader.c to support loading ELF binaries
(what the previous version already could do) and additional data modules
(something that the previous version could not do).

With this the kernel can obtain the list of reserved memory blocks from
UEFI loader. Each memory block is described by a name, begin and end
addresses.

UEFI loader will pass this way information about the memory occupied by
the kernel image itself, in this case the name will be "kernel".
Additionally this way the UEFI loader will pass to the kernel whatever
additional data modules are specified in the config.

Reserved memory doesn't include any memory the UEFI loader allocated to
store the name string and the array of the descriptors for reserved
memory blocks. So if the kernel for whatever reason wants to preserve
those it will have to copy the content somehow.